feat(product-service): stock decrement with optimistic and pessimistic locking
Ordering never decremented stock — you could order 1000 units of a product with 1
in stock — and the codebase had no locking of any kind.

The invariant lives in the entity: Product.decrementStock() refuses to go below
zero, so no caller can bypass it, backed by CHECK (stock_quantity >= 0) because
under concurrency only the database can guarantee it. Refusal maps to 409, not
400: the request was well-formed, the resource just isn't in a state that allows
it.

Two strategies, both implemented so they can be compared:

  optimistic  (V6) @Version — no lock, conflict DETECTED at write, needs a retry
  pessimistic      SELECT ... FOR UPDATE — conflict PREVENTED, needs no retry

The retry lives in a SEPARATE bean (StockService), not a private method. When an
optimistic conflict fires the transaction is already rollback-only, so the retry
must start a fresh one — and @Transactional is proxy-based, so a self-invocation
would open no transaction at all.

Retry uses exponential backoff with FULL JITTER. It started at 3 immediate
attempts and a 20-thread test proved that too low: only 6 of 20 succeeded,
because every loser re-tried in lockstep and collided with the same crowd.

Measured at 30-way contention: optimistic 649ms, pessimistic 116ms. Under low
contention the ranking reverses, since optimistic takes no lock at all.

Also completes the money migration's CONTRACT phase (V7 relax NOT NULL, V8 drop
the column — two steps, because reversing them breaks one code version or the
other), and opens the stock endpoints to any authenticated caller rather than
ADMIN: they are triggered by a customer ordering, not an admin editing.

Tests 44 -> 58: lost update reproduced at SQL level then shown rejected by
@Version, deadlock reproduced (SQLState 40P01) then made impossible by sorted
lock ordering, all four isolation anomalies, and a security regression test.

+++ product-service/src/test/java/org/example/productservice/product/infrastructure/ProductJpaRepositoryTest.java
@@ -33,7 +33,7 @@ class ProductJpaRepositoryTest extends AbstractIntegrationTest {
     private EntityManager em; // jakarta.persistence.EntityManager — to flush/clear the cache
 
     private Product newProduct(String name, double price, int stock) {
-        return new Product(UUID.randomUUID().toString(), name, price, stock);
+        return new Product(UUID.randomUUID().toString(), name, BigDecimal.valueOf(price), stock);
     }
 
     @Test
